chore: extract LogsReader and NotificationsWatcher to separate files
