+++ pkg/executionworker/interface.go
@@ -2,8 +2,6 @@ package executionworker
 
 import (
 	"context"
-	"io"
-	"sync/atomic"
 	"time"
 
 	testworkflowsv1 "github.com/kubeshop/testkube-operator/api/testworkflows/v1"
@@ -121,85 +119,6 @@ type NotificationsOptions struct {
 	ScheduledAt *time.Time // TODO: Consider no pointer
 	// NoFollow gives a hint to ignore following the further actions.
 	NoFollow bool
-}
-
-type notificationsWatcher struct {
-	ch       chan testkube.TestWorkflowExecutionNotification
-	finished atomic.Bool
-	err      atomic.Value
-}
-
-func newNotificationsWatcher() *notificationsWatcher {
-	return &notificationsWatcher{
-		ch: make(chan testkube.TestWorkflowExecutionNotification),
-	}
-}
-
-func (n *notificationsWatcher) send(notification testkube.TestWorkflowExecutionNotification) {
-	n.ch <- notification
-}
-
-func (n *notificationsWatcher) close(err error) {
-	if n.finished.CompareAndSwap(false, true) {
-		if err != nil {
-			n.err.Store(err)
-		}
-		close(n.ch)
-	}
-}
-
-func (n *notificationsWatcher) Channel() <-chan testkube.TestWorkflowExecutionNotification {
-	return n.ch
-}
-
-func (n *notificationsWatcher) Err() error {
-	err := n.err.Load()
-	if err == nil {
-		return nil
-	}
-	return err.(error)
-}
-
-type NotificationsWatcher interface {
-	Channel() <-chan testkube.TestWorkflowExecutionNotification
-	Err() error
-}
-
-type logsReader struct {
-	io.WriteCloser
-	io.Reader
-	finished atomic.Bool
-	err      atomic.Value
-}
-
-func newLogsReader() *logsReader {
-	reader, writer := io.Pipe()
-	return &logsReader{
-		Reader:      reader,
-		WriteCloser: writer,
-	}
-}
-
-func (n *logsReader) close(err error) {
-	if n.finished.CompareAndSwap(false, true) {
-		if err != nil {
-			n.err.Store(err)
-		}
-		n.WriteCloser.Close()
-	}
-}
-
-func (n *logsReader) Err() error {
-	err := n.err.Load()
-	if err == nil {
-		return nil
-	}
-	return err.(error)
-}
-
-type LogsReader interface {
-	io.Reader
-	Err() error
 }
 
 type Worker interface {
